- Updated changelog - Refactored storage - Fixed issue importing after logging out - Let Vite inject manifest link in html - Updated version
