+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.4] - 2023-10-12
+
+### Added
+
+### Changed
+- Refactoring storage
+
+### Fixed
+- Fixed issue importing previous data after logging out
+
+
 ## [0.3.3] - 2023-10-11
 
 ### Added
