+++ index.html
@@ -7,7 +7,6 @@
     <meta name="description" content="Record, track, and visualize your personal health data.">
     <meta name="theme-color" content="rgb(229, 231, 235)">
     <link rel="icon" href="/favicon.ico" sizes="any">
-    <link rel="manifest" href="/manifest.webmanifest">
     <link rel="apple-touch-icon" href="/assets/apple-icon-180.png">
     <meta name="apple-mobile-web-app-capable" content="yes">
     <link rel="apple-touch-startup-image" href="/assets/apple-splash-2048-2732.jpg" media="(device-width: 1024px) and (device-height: 1366px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)">
