+++ tailwind.config.js
@@ -4,9 +4,6 @@ export default {
     './index.html',
     './src/**/*.{vue,js,jsx,tsx}'
   ],
-  theme: {
-    extend: {},
-  },
   plugins: [
     require('@tailwindcss/forms')
   ],
